+++ index.html
@@ -230,6 +230,7 @@
   <!-- Экран загрузки изображения -->
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
+    <script type="module" src="js/main.js"></script>
   </template>
 
 </body>
