+++ js/main.js
@@ -1,47 +1,4 @@
-function randomNum(start, end) {
-  if (start > end) {
-    return -1;
-  } else if (start === end) {
-    return start;
-  }
-  return Math.floor(Math.random() * (end - start + 1)) + start;
-}
-
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const NAMES = ['Сергей', 'Мария', 'Иван', 'Аиша', 'Алексей', 'Артем', 'Полина', 'Дарья', 'Юрий'];
-
-const DESCRIPTIONS = ['Ну я', 'Я', 'Это я', 'Ну как я вам'];
-
-const createComment = {
-  id: randomNum(1, 100000000),
-  avatar:  `img/avatar-${randomNum(1, 6)}.svg`,
-  message: MESSAGES[randomNum(1, MESSAGES.length - 1)],
-  name: NAMES[randomNum(1, NAMES.length - 1)]
-};
-
-const createPost = (id) => ({
-  id: id,
-  url: `photos/${id}.jpg`,
-  description: DESCRIPTIONS[randomNum(1, DESCRIPTIONS.length - 1)],
-  likes: randomNum(15, 200),
-  comments: createComment
-});
-
-const generatePosts = () => {
-  const posts = [];
-  for (let i = 1; i <= 25; i++) {
-    posts.push(createPost(i));
-  }
-  return posts;
-};
+import {generatePosts} from './data';
 
 generatePosts();
